chore: update import paths and output paths to match new folder structure
