+++ project02/src/image_utils.py
